feat: add check for free disk space decompose upload handler logic
