+++ test/integration/upload.spec.ts
@@ -9,11 +9,11 @@ import Libp2p from 'libp2p'
 import path from 'path'
 import config from 'config'
 import PeerId, { JSONPeerId } from 'peer-id'
-import { rooms } from '../../src/communication'
 
+import { rooms } from '../../src/communication'
 import { MessageCodesEnum, ServiceAddresses, UploadJobStatus } from '../../src/definitions'
 import { ProviderManager } from '../../src/providers'
-import { UPLOAD_FOLDER } from '../../src/upload/handler/upload.handler'
+import { UPLOAD_FOLDER } from '../../src/upload/handler/upload'
 import UploadClient from '../../src/upload/model/clients.model'
 import UploadJob from '../../src/upload/model/upload.model'
 import { duplicateObject } from '../../src/utils'
